reorganize spotifyWrapper helper methods
